Improve the efficiency of batch model loads for indexing
When doing an initial index, we can use `find_in_batches` and `includes` to
get pretty efficient SQL queries. That was working for everything except our
audio, which was hitting the `self.audio.available` call and having to make a
new query each time we hit `to_article`. Changing that to a function improves
our index case and won't have any real-world impact on other uses.