Listen Live: Add a timeout to our preroll call
jQuery `error` handlers aren't called for JSONP requests, so replace our
error handling there with a timeout that makes sure we start the stream if
our preroll request hasn't succeeded within 2.5 seconds.

If we error, this means a 2.5 second delay before the stream starts.

If the request is just super-slow, we would still show the visual but
wouldn't play the preroll and wouldn't count the impression.