Expire fragment cache on segments more intelligently
Previously, segment fragment caches were getting expired only when
the segment itself was updated. Consequently, the "more from this
episode" feature would not appear on segments if a segment was
added to an episode after the segment was published or updated.

Adding @segment.published_episode to the fragment cache gives us
the ability to expire also when the segment's episode is updated.