Overload how IceCube handles dates with zones
IceCube, the gem we use for scheduling, serializes `start_date` into
`RecurringScheduleRule.schedule_hash`. For times with zones, it creates a
hash with the time and zone. That breaks our ES indexing, but we've never
noticed because a) we haven't been passing it zones in production and
b) the exceptions were in `after_commit`, so they were getting swallowed up
in tests.

Overload the serialization to instead use iso8601 datetime strings with
offset.